Underneath the introductory comments, you should add the commit message to the text file.
To write a useful commit message, you should include a summary on the
first line that is around 50 characters long. Under this, and broken
up into digestible sections, you should include a description that
states the reason you made this change, how the code works, and
additional information that will contextualize and clarify it for
others to review the work when merging it. Try to be as helpful and
proactive as possible to ensure that those maintaining the project
are able to fully understand your contribution.